move enums from Gapic to Enums dir
