+++ tests/Integration/goldens/compute_small/src/V1/Enums/Address/AddressType.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2022 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,21 +18,23 @@
 /*
  * GENERATED CODE WARNING
  * Generated by gapic-generator-php from the file
- * https://github.com/googleapis/googleapis/blob/master/tests/Unit/ProtoTests/BasicDiregapic/library_rest.proto
+ * https://github.com/googleapis/googleapis/blob/master/tests/Integration/apis/compute_small/v1/compute_small.proto
  * Updates to the above are reflected here through a refresh process.
  */
 
-namespace Testing\BasicDiregapic\Gapic\SomeMessage;
+namespace Google\Cloud\Compute\V1\Enums\Address;
 
 /**
- * Alignment contains string constants that represent the names of each value in
- * the google.example.library.v1.SomeMessage.Alignment descriptor.
+ * AddressType contains string constants that represent the names of each value in
+ * the google.cloud.compute.v1.Address.AddressType descriptor.
  */
-class Alignment
+class AddressType
 {
-    const GOOD = 'GOOD';
+    const UNDEFINED_ADDRESS_TYPE = 'UNDEFINED_ADDRESS_TYPE';
 
-    const NEUTRAL = 'NEUTRAL';
+    const EXTERNAL = 'EXTERNAL';
 
-    const EVIL = 'EVIL';
+    const INTERNAL = 'INTERNAL';
+
+    const UNSPECIFIED_TYPE = 'UNSPECIFIED_TYPE';
 }
