+++ tests/Integration/goldens/compute_small/src/V1/Enums/Address/NetworkTier.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2022 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,23 +18,21 @@
 /*
  * GENERATED CODE WARNING
  * Generated by gapic-generator-php from the file
- * https://github.com/googleapis/googleapis/blob/master/tests/Unit/ProtoTests/BasicDiregapic/library_rest.proto
+ * https://github.com/googleapis/googleapis/blob/master/tests/Integration/apis/compute_small/v1/compute_small.proto
  * Updates to the above are reflected here through a refresh process.
  */
 
-namespace Testing\BasicDiregapic\Gapic\Comment;
+namespace Google\Cloud\Compute\V1\Enums\Address;
 
 /**
- * Stage contains string constants that represent the names of each value in the
- * google.example.library.v1.Comment.Stage descriptor.
+ * NetworkTier contains string constants that represent the names of each value in
+ * the google.cloud.compute.v1.Address.NetworkTier descriptor.
  */
-class Stage
+class NetworkTier
 {
-    const UNSET = 'UNSET';
+    const UNDEFINED_NETWORK_TIER = 'UNDEFINED_NETWORK_TIER';
 
-    const DRAFT = 'DRAFT';
+    const PREMIUM = 'PREMIUM';
 
-    const PUBLISHED = 'PUBLISHED';
-
-    const DELETED = 'DELETED';
+    const STANDARD = 'STANDARD';
 }
