+++ tests/Integration/goldens/compute_small/src/V1/Enums/Address/Purpose.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2022 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,19 +18,25 @@
 /*
  * GENERATED CODE WARNING
  * Generated by gapic-generator-php from the file
- * https://github.com/googleapis/googleapis/blob/master/tests/Unit/ProtoTests/BasicDiregapic/library_rest.proto
+ * https://github.com/googleapis/googleapis/blob/master/tests/Integration/apis/compute_small/v1/compute_small.proto
  * Updates to the above are reflected here through a refresh process.
  */
 
-namespace Testing\BasicDiregapic\Gapic\BookResponse;
+namespace Google\Cloud\Compute\V1\Enums\Address;
 
 /**
- * Rating contains string constants that represent the names of each value in the
- * google.example.library.v1.BookResponse.Rating descriptor.
+ * Purpose contains string constants that represent the names of each value in the
+ * google.cloud.compute.v1.Address.Purpose descriptor.
  */
-class Rating
+class Purpose
 {
-    const GOOD = 'GOOD';
+    const UNDEFINED_PURPOSE = 'UNDEFINED_PURPOSE';
 
-    const BAD = 'BAD';
+    const DNS_RESOLVER = 'DNS_RESOLVER';
+
+    const GCE_ENDPOINT = 'GCE_ENDPOINT';
+
+    const NAT_AUTO = 'NAT_AUTO';
+
+    const VPC_PEERING = 'VPC_PEERING';
 }
