+++ tests/Integration/goldens/compute_small/src/V1/Enums/Address/Status.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2022 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,19 +18,23 @@
 /*
  * GENERATED CODE WARNING
  * Generated by gapic-generator-php from the file
- * https://github.com/googleapis/googleapis/blob/master/tests/Unit/ProtoTests/BasicDiregapic/library_rest.proto
+ * https://github.com/googleapis/googleapis/blob/master/tests/Integration/apis/compute_small/v1/compute_small.proto
  * Updates to the above are reflected here through a refresh process.
  */
 
-namespace Testing\BasicDiregapic\Gapic;
+namespace Google\Cloud\Compute\V1\Enums\Address;
 
 /**
- * TopLevelEnum contains string constants that represent the names of each value in
- * the google.example.library.v1.TopLevelEnum descriptor.
+ * Status contains string constants that represent the names of each value in the
+ * google.cloud.compute.v1.Address.Status descriptor.
  */
-class TopLevelEnum
+class Status
 {
-    const FOO = 'FOO';
+    const UNDEFINED_STATUS = 'UNDEFINED_STATUS';
 
-    const BAR = 'BAR';
+    const IN_USE = 'IN_USE';
+
+    const RESERVED = 'RESERVED';
+
+    const RESERVING = 'RESERVING';
 }
