+++ tests/Integration/goldens/compute_small/src/V1/Enums/Warning/Code.php
@@ -22,11 +22,11 @@
  * Updates to the above are reflected here through a refresh process.
  */
 
-namespace Google\Cloud\Compute\V1\Gapic\Warnings;
+namespace Google\Cloud\Compute\V1\Enums\Warning;
 
 /**
  * Code contains string constants that represent the names of each value in the
- * google.cloud.compute.v1.Warnings.Code descriptor.
+ * google.cloud.compute.v1.Warning.Code descriptor.
  */
 class Code
 {
